refactor(reborn): add producer-facing Resolution constructors in ironclaw_turns (§5.3 Stage 2b)
New run_profile::resolution module with completed/failed/spawned_process/
spawned_child_run (return Resolution), approval_required/auth_required/
resource_blocked/await_dependent_run/external_tool_pending (return
GatedResolution), and denied (returns DeniedResolution). Moves the non-lossy
redaction verbatim from resolution_mapping (ModelResultPreview 24 KiB
word-boundary redaction, ModelFailureDiagnostic, deterministic
GateRef::for_auth_gate auth key, resume-token carry, DependentRunResult inline
staged result, DenyReason mapping). resolution_mapping::
capability_outcome_to_resolution is retained transitionally as a thin delegator
so unmigrated producers keep compiling; RefBindings is now empty (loop refs ride
the channel origin). 17 retargeted constructor tests green.
